Add _data/site.json and update site-header.html
Creates the new /src/_data directory with the new site.json file in it. Adds a json object holding the site name and url in the site.json file. Updates the /src/_includes/partials/site-header.html with template tags, attached to an ARIA label, to pull in the site name. This adds the site name to the site header.