Create navigation.json and use Nunjucks for links
Adds /src/_data/navigation.json file with an array of objects that link to
top-level pages. Replaces standard <ul> nav list in the site-header
file with a Nunjucks loop and template tags. Uses Nunjucks to loop through each item (from the navigation.json file) and create a link with it.