Add stylesheet link to base template
The stylesheet link is formed with Nunjucks template tags. It is not
necessary to write the link in this way. It could be written in the
usual way as a straightforward HTML tag. At this stage I do not know if
including it as Nunjucks templating brings any benefit. It links to the
styles.css file that will be transpiled from the Bulma styles.scss file.