Add gulpfile.js for storing Gulp configurations
The gulpfile.js is the essential configuration file for working with
Gulp, a task runner tool. This initial commit of the file contains
working code for the first Gulp task in this project. That task uses
Sass to preprocess an SCSS file styles.scss and output CSS to styles.css.
The CSS output is compressed. The styles.scss file is the working file
to customise the Bulma framework, which is the CSS framework being used
in the project.