Add a directory to hold documents for download
We sometimes have to put documents on our website for people to
download. For example, job packs for job vacancies. It is good practice
to limit such documents; however, we are not in a position to illiminate
this requirement. At this stage, these documents will be stored in a
directory in our repository. This commit adds that directory and two
test documents. I linked to them (to download) from a news-item post.