Make a news collection and passthrough documents
Sets a global filter to group news item posts. The news items are in a
news directory. The filter takes all the Markdown files in that
directory. The results are sorted using an order number (which is set in
the front matter of each news item). This allows control over the order
in which items appear on a page.

Separately adds a passthrough for the documents directory. The documents
directory was previously added to create a store for downloadable
documents. Examples of such documents are: job packs for vacancies and
policy documents.